Create GameOfLife.pl
Adds a prolog version of the game of life.

I kept this impl after the 2014 Global Day of Code Retreat.  For the last session, I banged this out in about 30 minutes, then struggled with swi-pl for the next 15, because I couldn't get it to render the 'nl' characters.

To use, start up prolog; import this file (something like "['GameOfLife.pl']."); seed the field with some variety of "createCell(XCoord,YCoord)."; and then use the commands "tick" and print(Xmin, YMin, XMax, YMax) to increment the generation and print the field.  

The consequence of the "nl" characters not printing is that the grid prints as a single line.  Oh well.  Not bad for a half an hour.